Switch v1_20_R3 -> v1_20_R4
