+++ core/src/main/java/be/isach/ultracosmetics/version/ServerVersion.java
@@ -8,19 +8,11 @@ import org.bukkit.Bukkit;
 public enum ServerVersion {
 
     // Do not supply a mapping version when there is no NMS module.
-    v1_8(8, 8),
-    v1_9(9, 4),
-    v1_10(10, 2),
-    v1_11(11, 2),
-    v1_12(12, 2),
-    v1_13(13, 2),
-    v1_14(14, 4),
-    v1_15(15, 2),
-    v1_16(16, 5),
+    // 1.17 is the first version to support Java 17
     v1_17(17, 1),
     v1_18(18, 2),
     v1_19(19, 4, "3009edc0fff87fa34680686663bd59df", 3),
-    v1_20(20, 4, "60a2bb6bf2684dc61c56b90d7c41bddc", 3),
+    v1_20(20, 6, "ee13f98a43b9c5abffdcc0bb24154460", 4),
     NEW("???"),
     ;
