+++ settings.gradle
@@ -1,2 +1,2 @@
-include 'v1_19_R3', 'v1_20_R3', 'core', 'worldguard', 'worldguard-legacy', 'java-17'
+include 'v1_19_R3', 'v1_20_R4', 'core', 'worldguard', 'worldguard-legacy', 'java-17'
 rootProject.name = 'UltraCosmetics'
