--- v1_20_R3/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    implementation 'org.spigotmc:spigot:1.20.4-R0.1-SNAPSHOT:remapped-mojang'
-    implementation project(':core')
-}
