--- v1_20_R3/gradle.properties
@@ -1,2 +1,0 @@
-javaVersion=17
-remapMcVersion=1.20.4
