--- v1_20_R3/src/main/java/be/isach/ultracosmetics/v1_20_R3/EntityBase.java
@@ -1,8 +1,0 @@
-package be.isach.ultracosmetics.v1_20_R3;
-
-/**
- * @author RadBuilder
- */
-public interface EntityBase {
-    public void travel_(float sideMot, float forMot);
-}
