--- v1_20_R3/src/main/java/be/isach/ultracosmetics/v1_20_R3/ObfuscatedFields.java
@@ -1,16 +1,0 @@
-package be.isach.ultracosmetics.v1_20_R3;
-
-/**
- * This file automatically generated by obfuscatedfields_generator.py in tooling folder
- */
-public class ObfuscatedFields {
-    // corresponds to net.minecraft.world.entity.LivingEntity#jumping
-    public static final String JUMPING = "bj";
-    // corresponds to net.minecraft.core.MappedRegistry#frozen
-    public static final String FROZEN = "l";
-    // corresponds to net.minecraft.core.MappedRegistry#unregisteredIntrusiveHolders
-    public static final String INTRUSIVE_HOLDER_CACHE = "m";
-
-    private ObfuscatedFields() {
-    }
-}
