--- v1_20_R3/src/main/java/be/isach/ultracosmetics/v1_20_R3/mount/MountSlime.java
@@ -1,28 +1,0 @@
-package be.isach.ultracosmetics.v1_20_R3.mount;
-
-import be.isach.ultracosmetics.UltraCosmetics;
-import be.isach.ultracosmetics.cosmetics.type.MountType;
-import be.isach.ultracosmetics.player.UltraPlayer;
-import be.isach.ultracosmetics.v1_20_R3.customentities.CustomSlime;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
-import org.bukkit.craftbukkit.v1_20_R3.entity.CraftPlayer;
-
-/**
- * @author RadBuilder
- */
-public class MountSlime extends MountCustomEntity {
-
-    public MountSlime(UltraPlayer owner, MountType type, UltraCosmetics ultraCosmetics) {
-        super(owner, type, ultraCosmetics);
-    }
-
-    @Override
-    public void onUpdate() {
-    }
-
-    @Override
-    public LivingEntity getNewEntity() {
-        return new CustomSlime(EntityType.SLIME, ((CraftPlayer) getPlayer()).getHandle().level());
-    }
-}
