+++ v1_20_R4/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation 'org.spigotmc:spigot:1.20.6-R0.1-SNAPSHOT:remapped-mojang'
+    implementation project(':core')
+}
