+++ v1_20_R4/gradle.properties
@@ -1,0 +1,2 @@
+javaVersion=21
+remapMcVersion=1.20.6
