+++ v1_20_R4/src/main/java/be/isach/ultracosmetics/v1_20_R4/EntityBase.java
@@ -1,0 +1,8 @@
+package be.isach.ultracosmetics.v1_20_R4;
+
+/**
+ * @author RadBuilder
+ */
+public interface EntityBase {
+    public void travel_(float sideMot, float forMot);
+}
